Remove parens from calls to `time`
In Perl6 `time` returns the number of seconds since epoch, which is what the
`time()` function does in Perl5 and hence what is required here.  When one
uses parens on `time`, the compiler tries to look for a `time()` routine in
the current scope and can't find it, hence this change allows the tests to
run a bit further.